--- dcore/__init__.py
@@ -1,8 +1,0 @@
-from __future__ import absolute_import
-from dcore.configuration import *
-from dcore.advection import *   # noqa
-from dcore.diagnostics import *     # noqa
-from dcore.forcing import *     # noqa
-from dcore.linear_solvers import *  # noqa
-from dcore.state import *           # noqa
-from dcore.timeloop import *        # noqa
