+++ examples/dcmip_3-1_linear_advection_quads_rt0.py
@@ -1,4 +1,4 @@
-from dcore import *
+from gusto import *
 from firedrake import CubedSphereMesh, ExtrudedMesh, Expression, \
     VectorFunctionSpace
 from firedrake import exp, acos, cos, sin
