+++ examples/embedded_DG.py
@@ -1,4 +1,4 @@
-from dcore import *
+from gusto import *
 from firedrake import IcosahedralSphereMesh, Expression, SpatialCoordinate, \
     as_vector, VectorFunctionSpace, File
 import itertools
