+++ gusto/__init__.py
@@ -1,0 +1,8 @@
+from __future__ import absolute_import
+from gusto.configuration import *
+from gusto.advection import *   # noqa
+from gusto.diagnostics import *     # noqa
+from gusto.forcing import *     # noqa
+from gusto.linear_solvers import *  # noqa
+from gusto.state import *           # noqa
+from gusto.timeloop import *        # noqa
